Bump version to 1.3.1 and update release notes
Updated the version number in multiple files from `01.03.00` to `01.03.01`, indicating a new release of the Power BI Embedded modules. Changed the version displayed in the `ReleaseBadge.svg` from `v1.3.0` to `v1.3.1`. Added a new entry in `ReleaseNotes.txt` for version `1.3.1`, which includes a fix for downloading PBIX files for LiveConnect reports. Modified the `DotNetNuke.PowerBI.dnn` file to reflect the updated version numbers for the `PowerBIEmbeddedList`, `PowerBIEmbeddedContent`, and `PowerBIEmbeddedCalendar` packages. Updated the assembly version in `AssemblyInfo.cs` to `01.03.01.*`, ensuring that the build and revision numbers are automatically generated.